Enabled continuation calculation of He_ph matrix ONLY when using the auxillary files. This has two consequences: 1) One does not need to redo all calculations again and again 2) Inelastica can check whether all He_ph couplings are calculated
Note that if not using the auxillary file one cannot continue on the calculation.
Also note that the new global attribute CurrentHWidx is used to track this.
To restart a calculation one should delete the file.
I have checked this with the AuH2 example and both checked the output of the PHrun and the dI curve from Inelastica