Removed print statements from installation
Print statements when installing does not help (at all).
Hence they have been removed.

The check for slow BLAS/LAPACK still remains, but now it
does not force the user to press a key. 
Say if any-body wishes to use an automatic script to install
this _and_ have a slow BLAS, then this is not ideal.
Besides, any slow performance in Inelastica/BLAS/LAPACK will be due 
to the numpy installation.

Instead of doing sys.exit, better to raise an Error, for this the
NameError provides apt usage as also suggested by the python community.