tests: added a simple 1D carbon chain system
This test system comes in three different representations, namely
with the chain oriented along the first, second, or third lattice
vector, i.e., the x-, y-, or z-axis. The computed physical
quantities should thus be identical.

Before executing `python script.py` one needs to first run `transiesta`
in the ELEC and OSrun directories, and thereafter in the TSrun and FCrun_9.